removed type explicit cast again, using it in assembly was (obviously) illegal

+++ include/constants/forms.h
@@ -5,7 +5,7 @@
 
 #define FORM_FLAG_SHIFT 14
 // zero form is simply the 'normal' Gen 3 pokemon!
-#define FORM_FLAG_VALUE(form) ((u16)(form) << FORM_FLAG_SHIFT)
+#define FORM_FLAG_VALUE(form) ((form) << FORM_FLAG_SHIFT)
 #define FIRST_FORM FORM_FLAG_VALUE(1)
 #define SECOND_FORM FORM_FLAG_VALUE(2)
 #define THIRD_FORM FORM_FLAG_VALUE(3)
